Add `lotus-lib`, migrate to ESM mode (#8)
This commit adds the new `lotus-lib` submodule and installs it. There
are also changes made to the package.json, TypeScript, and NPM
configuration to ensure proper transpilation and dependency install.

All of the required changes to imports were made. Also, Prisma was
updated and the schema was updated to generate ESM modules instead of
CommonJS (i.e. `prisma-client` vs `prisma-client-js` provider).

Tested `lotus-lib` integration with the beta bot and confirmed Gives and
Withdrawals are crafting and broadcasting valid transactions.