Revert back to CommonJS mode (#9)
This commit reverts back to CommonJS mode for both `lotus-lib` and `lotus-bot` configurations. The purpose for this change is to fix a regression with the Temporal workflow bundles. In ESM mode, the workflow bundles cannot be properly loaded.

The issue with Temporal workflow bundles can most likely be fixed in Temporal's inline Webpack configuration hook, but we can fix that later.